router로 block 안에 화면 바뀌는거 구현하고 싶은데 잘 안되네 .. #6
