+++ src/pages/MusicList.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function MusicList() {
+  return <div>musicList</div>;
+}
